[R2DBC-67] set SPEC version support tp 1.0.0-release version

+++ src/benchmark/java/org/mariadb/r2dbc/Do_1.java
@@ -9,16 +9,16 @@ import reactor.core.publisher.Flux;
 public class Do_1 extends Common {
 
   @Benchmark
-  public Integer testR2dbc(MyState state) throws Throwable {
+  public Long testR2dbc(MyState state) throws Throwable {
     return consume(state.r2dbc);
   }
 
   @Benchmark
-  public Integer testR2dbcPrepare(MyState state) throws Throwable {
+  public Long testR2dbcPrepare(MyState state) throws Throwable {
     return consume(state.r2dbcPrepare);
   }
 
-  private Integer consume(io.r2dbc.spi.Connection connection) {
+  private Long consume(io.r2dbc.spi.Connection connection) {
     io.r2dbc.spi.Statement statement = connection.createStatement("DO 1");
     return
         Flux.from(statement.execute())
